COM-3408 - define getPotentialTrainees in attendance composable
